requirements.txt
a2wsgi==1.10.10
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
aiosqlite==0.21.0
alembic==1.14.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.7.0
apache-airflow==2.10.4
apache-airflow-core==3.0.4
apache-airflow-providers-celery==3.8.5
apache-airflow-providers-common-compat==1.2.2
apache-airflow-providers-common-io==1.4.2
apache-airflow-providers-common-sql==1.28.1
apache-airflow-providers-fab==1.5.1
apache-airflow-providers-ftp==3.11.1
apache-airflow-providers-http==4.13.3
apache-airflow-providers-imap==3.7.0
apache-airflow-providers-microsoft-mssql==4.3.2
apache-airflow-providers-postgres==5.14.0
apache-airflow-providers-redis==3.8.0
apache-airflow-providers-smtp==1.8.1
apache-airflow-providers-sqlite==3.9.1
apache-airflow-providers-standard==1.5.0
apache-airflow-task-sdk==1.0.4
apispec==6.8.0
argcomplete==3.5.2
asgiref==3.8.1
asyncpg==0.30.0
attrs==24.2.0
babel==2.16.0
billiard==4.2.1
blinker==1.9.0
cachelib==0.9.0
cadwyn==5.4.4
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
colorama==0.4.6
colorlog==6.9.0
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.5
croniter==5.0.1
cryptography==42.0.8
Deprecated==1.2.15
dill==0.3.1.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
Flask==2.2.5
Flask-AppBuilder==4.5.2
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.9.2
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
flower==2.0.1
frozenlist==1.5.0
fsspec==2024.10.0
google-re2==1.1.20240702
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.0
humanize==4.11.0
idna==3.10
importlib-metadata==6.11.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.2
lazy-object-proxy==1.10.0
libcst==1.8.2
limits==3.14.1
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.5.0
msgspec==0.19.0
multidict==6.1.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
ordered-set==4.1.0
packaging==24.2
pathspec==0.12.1
pendulum==3.0.0
pluggy==1.5.0
prison==0.2.1
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==4.25.5
psutil==6.1.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.18.0
PyJWT==2.10.1
pymssql==2.3.8
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
retryhttp==1.3.3
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.6.0
rich-toolkit==0.14.9
rignore==0.6.4
rpds-py==0.22.3
sentry-sdk==2.34.1
setproctitle==1.3.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
starlette==0.47.2
structlog==25.4.0
svcs==25.1.0
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.5.0
text-unidecode==1.3
time-machine==2.16.0
tornado==6.4.2
typer==0.16.0
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
Unidecode==1.4.0
universal_pathlib==0.2.5
urllib3==2.2.3
uuid6==2025.0.1
uvicorn==0.35.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==2.2.3
wirerope==0.4.8
wrapt==1.17.0
WTForms==3.2.1
yarl==1.18.3
zipp==3.21.0
